# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StudyBuddy OnDemand API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/content/{unit_id}/lesson": {
      "get": {
        "tags": ["content"],
        "summary": "Get Lesson",
        "description": "Serve a lesson for the given unit.\n\nEntitlement guard: free-tier students are limited to 2 lessons.\nIncrements lessons_accessed after serving.",
        "operationId": "get_lesson_api_v1_content__unit_id__lesson_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/lesson/audio": {
      "get": {
        "tags": ["content"],
        "summary": "Get Lesson Audio",
        "description": "Return a URL for the lesson MP3 audio file.\n\nFor local dev (no S3): returns a /static/content/... URL.\nFor production: returns a pre-signed S3 URL.\n\nNever proxies audio bytes through this server.",
        "operationId": "get_lesson_audio_api_v1_content__unit_id__lesson_audio_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/quiz": {
      "get": {
        "tags": ["content"],
        "summary": "Get Quiz",
        "description": "Serve a quiz set, rotating through sets 1\u21922\u21923\u21921 per student per unit.",
        "operationId": "get_quiz_api_v1_content__unit_id__quiz_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/tutorial": {
      "get": {
        "tags": ["content"],
        "summary": "Get Tutorial",
        "description": "Serve the tutorial for a unit.",
        "operationId": "get_tutorial_api_v1_content__unit_id__tutorial_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorialResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/experiment": {
      "get": {
        "tags": ["content"],
        "summary": "Get Experiment",
        "description": "Serve the lab experiment for a unit.\nReturns 404 if no experiment file exists (non-lab unit).",
        "operationId": "get_experiment_api_v1_content__unit_id__experiment_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/report": {
      "post": {
        "tags": ["content"],
        "summary": "Report Content",
        "description": "Student reports a content issue.\nStores in student_content_feedback table.",
        "operationId": "report_content_api_v1_content__unit_id__report_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/content/{unit_id}/feedback/marked": {
      "post": {
        "tags": ["content"],
        "summary": "Submit Marked Feedback",
        "description": "Student submits feedback on a specific marked portion of content.\nStored as student_content_feedback with content_type from body.",
        "operationId": "submit_marked_feedback_api_v1_content__unit_id__feedback_marked_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": ["url", "expires_in"],
        "title": "AudioUrlResponse"
      },
...
      "ExperimentQuestion": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": ["question", "answer"],
        "title": "ExperimentQuestion"
      },
      "ExperimentResponse": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "experiment_title": {
            "type": "string",
            "title": "Experiment Title"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Materials"
          },
          "safety_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Safety Notes"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/ExperimentStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ExperimentQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "conclusion_prompt": {
            "type": "string",
            "title": "Conclusion Prompt"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "content_version": {
            "type": "integer",
            "title": "Content Version"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "language",
          "experiment_title",
          "materials",
          "safety_notes",
          "steps",
          "questions",
          "conclusion_prompt",
          "generated_at",
          "model",
          "content_version"
        ],
        "title": "ExperimentResponse"
      },
...
      "FeedbackRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "start_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "end_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          },
          "marked_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marked Text"
          },
          "feedback_text": {
            "type": "string",
            "title": "Feedback Text"
          }
        },
        "type": "object",
        "required": ["content_type", "feedback_text"],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonResponse": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "grade": {
            "type": "integer",
            "title": "Grade"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/LessonSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Points"
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio",
            "default": false
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "content_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Version"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "title",
          "grade",
          "subject",
          "lang",
          "sections",
          "key_points"
        ],
        "title": "LessonResponse"
      },
      "LessonSection": {
        "properties": {
          "heading": {
            "type": "string",
            "title": "Heading"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": ["heading", "body"],
        "title": "LessonSection"
      },
...
      "QuizQuestion": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text"
          },
          "question_type": {
            "type": "string",
            "title": "Question Type"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/QuizOption"
            },
            "type": "array",
            "title": "Options"
          },
          "correct_option": {
            "type": "string",
            "title": "Correct Option"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "question_text",
          "question_type",
          "options",
          "correct_option",
          "explanation",
          "difficulty"
        ],
        "title": "QuizQuestion"
      },
      "QuizResponse": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "set_number": {
            "type": "integer",
            "title": "Set Number"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "estimated_duration_minutes": {
            "type": "integer",
            "title": "Estimated Duration Minutes"
          },
          "passing_score": {
            "type": "integer",
            "title": "Passing Score"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "content_version": {
            "type": "integer",
            "title": "Content Version"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "set_number",
          "language",
          "questions",
          "total_questions",
          "estimated_duration_minutes",
          "passing_score",
          "generated_at",
          "model",
          "content_version"
        ],
        "title": "QuizResponse"
      },
...
      "ReportRequest": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["category"],
        "title": "ReportRequest"
      },
...
      "TutorialResponse": {
        "properties": {
          "unit_id": {
            "type": "string",
            "title": "Unit Id"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/TutorialSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "common_mistakes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Common Mistakes"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "content_version": {
            "type": "integer",
            "title": "Content Version"
          }
        },
        "type": "object",
        "required": [
          "unit_id",
          "language",
          "title",
          "sections",
          "common_mistakes",
          "generated_at",
          "model",
          "content_version"
        ],
        "title": "TutorialResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}